Adding the current progress
I did not think about having this onto git hub until a few weeks after I started so here is what I have done.

Tried making a visualizer in the UI toolkit, and with the legacy editor UI. The UI Toolkit had a visual bug so I gave up. I gave up on the legacy UI because I had not even made a State Machine. for it to work with.

Now I am making a easy to use state machine that should work in any project.