Fixed Saving. Fixed Adding new Children braking UI
I made it so it will actually save the the state and the children. I used serializedObject.ApplyModifiedProperties, AssetDatabase.SaveAssets, and EditorUtility.SetDirty to make sure it saved. I then fixed a problem with the UI when you added a new child , it would copy the UI of the child at index 0.